Optimization for the Nodes' VLA.
As the length of the VLA is most often known at compile-time, having a 
field to record it was wasteful. That field is now merged into the VLA 
as an `uintptr_t` at the first index (must be cast manually). For the 
most common case where the length is constant, a table has been built to 
record it.