A nice vanilia bf machine — very fragile…
It crashes if it encounters anything but `+-.,<>[]`, or if it `<` at
the first cell. It does allow incrementing above 0xff, or decrementing
under 0, which I dont want to.
This is for the sake of simplicity, and also because I plan to use it as
bytecode VM later, so it will always be given clean code anyway.

The memory, however, is managed, and the tape can grow (on the right) as
needed.